pyproject.toml
[project]
name = "video_sampler"
description = "Video Sampler -- sample frames from a video file"
url = "https://github.com/LemurPwned/video-sampler"
version = "0.11.2"
authors = [
    { name = "LemurPwned", email = "lemurpwned@gmail.com" }
]
readme = "README.md"
requires-python = ">3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]

keywords = [
  "video sampling",
  "frame selection",
  "labelling",
  "labeling",
  "annotation",
]

dependencies = [
  "av >=10.0.0",
  "ImageHash >=4.3.1",
  "Pillow >=10.0.0",
  "rich >= 13.5.3",
  "typer >= 0.12.5",
  "tqdm >= 4.66.1",
  "opencv-python-headless >= 4.9.0.80",
  "pydantic >= 2.6.1",
  "pyaml >= 6.0.1",
  "requests >= 2.31.0",
]

[project.urls]
Source = "https://github.com/LemurPwned/video-sampler"

[project.scripts]
video_sampler = "video_sampler.__main__:main_loop"

[project.optional-dependencies]
clip = [
  "open_clip_torch >= 2.23.0",
  "torch >= 2.1.0"
]
ytdlp = [
  "yt-dlp >= 2023.12.30"
]
language = [
  "spacy >= 3.7.0",
  "pysrt >= 1.1.2"
]
dev = [
    "pre-commit",
    "bump2version",
    "tabulate",
    "pytest"
]
all = [
  "open_clip_torch >= 2.23.0",
  "torch >= 2.1.0",
  "yt-dlp >= 2023.12.30",
  "spacy >= 3.7.0",
  "pysrt >= 1.1.2",
  "openai >= 1.16.2",
]

[build-system]
build-backend = "flit_core.buildapi"
requires = ["flit_core >=3.2,<4"]
